+++ example/components/Toolbar.styles.js
@@ -1,0 +1,30 @@
+import { StyleSheet } from 'react-native'
+
+export const createToolbarStyles = (theme) => {
+	return StyleSheet.create({
+		toolbar: {
+			flexDirection: 'row',
+			justifyContent: 'space-between',
+			alignItems: 'center',
+			paddingHorizontal: 16,
+			paddingVertical: 12,
+			backgroundColor: theme.headerColor,
+			borderBottomWidth: 1,
+			borderBottomColor: theme.gridColor,
+		},
+		title: {
+			fontSize: 18,
+			fontWeight: 'bold',
+			color: theme.textColor,
+		},
+		toolbarRight: {
+			flexDirection: 'row',
+			alignItems: 'center',
+		},
+		themeLabel: {
+			fontSize: 14,
+			color: theme.textColor,
+			marginRight: 8,
+		},
+	})
+}
